Refactor and simplify code
